middleware

+++ app/Http/Controllers/TweetController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Tweet;
+use App\Tweet; // Need to pull in our model to use it!
+use App\User; // Let's pull in our User model!
+use Auth; // Need to pull in Auth in order to use it!
 
 class TweetController extends Controller
 {
@@ -15,9 +17,11 @@ class TweetController extends Controller
     public function index()
     {
         //
-        $tweets = Tweet::all();
+        $tweets = Tweet::query()
+                    ->join( 'users', 'tweets.user_id', '=', 'users.id' )
+                    ->get();
 
-        Return view('tweets.index', compact('tweets'));
+        return view('tweets.index', compact('tweets'));
     }
 
     /**
@@ -28,7 +32,11 @@ class TweetController extends Controller
     public function create()
     {
         //
-        return view('tweets.create');
+        $user = Auth::user();
+        if ( $user ) // Yay! You're logged in, create away!
+            return view('tweets.create');
+        else // Uh oh, logged out! Redirect.
+            return redirect('/tweets');
     }
 
     /**
@@ -39,15 +47,20 @@ class TweetController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        $validatedData = $request->validate(array(
-          'message' => 'required|max:255',
-          'author'  => 'required|max:64'
-          ));
+        // Assign and check user all at once.
+        if ( $user = Auth::user() ) { // Proceed and store data if the user is logged in.
+            $validatedData = $request->validate(array(
+                'message' => 'required|max:255'
+            ));
+            $tweet = new Tweet;
+            $tweet->user_id = $user->id;
+            $tweet->message = $validatedData['message'];
+            $tweet->save();
 
-          $tweet = Tweet::create($validatedData);
-
-          return redirect('/tweets')->with('success', 'Tweet saved.');
+            return redirect('/tweets')->with('success', 'Tweet saved.');
+        }
+        // Redirect by default.
+        return redirect('/tweets');
     }
 
     /**
@@ -59,6 +72,12 @@ class TweetController extends Controller
     public function show($id)
     {
         //
+        $tweet = Tweet::findOrFail($id);
+              $userID = $tweet->user_id;
+              $tweetUser = User::query()->where('users.id','=', $userID)->get()[0];
+        //$tweetUser = $tweet->user()->get()[0];
+
+        return view( 'tweets.show', compact('tweet'), compact('tweetUser') );
     }
 
     /**
@@ -69,10 +88,13 @@ class TweetController extends Controller
      */
     public function edit($id)
     {
-        //
-        $tweet = Tweet::findOrFail($id);
-
-        return view('tweets.edit', compact('tweet'));
+        // Check if user is logged in.
+        if ( $user = Auth::user() ) {
+            $tweet = Tweet::findOrFail($id);
+            return view( 'tweets.edit', compact('tweet') );
+        }
+        // Redirect by default.
+        return redirect('/tweets');
     }
 
     /**
@@ -84,16 +106,18 @@ class TweetController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
-        $validatedData = $request->validate(array(
-          'message' => 'required|max:255',
-          'author'  => 'required|max:64'
-          ));
+        // Check if user is logged in.
+        if ( $user = Auth::user() ) {
+            $validatedData = $request->validate(array(
+                'message' => 'required|max:255'
+            ));
 
-          Tweet::whereId($id)->update($validatedData);
+            Tweet::whereId($id)->update($validatedData);
 
-          return redirect('/tweets')->with('success', 'Tweet updated.');
-
+            return redirect('/tweets')->with('success', 'Tweet updated.');
+        }
+        // Redirect by default.
+        return redirect('/tweets');
     }
 
     /**
@@ -104,11 +128,15 @@ class TweetController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $tweet = Tweet::findOrFail($id);
+        // Check if user is logged in.
+        if ( $user = Auth::user() ) {
+            $tweet = Tweet::findOrFail($id);
 
-        $tweet->delete();
+            $tweet->delete();
 
-        return redirect('/tweets')->with('success', 'Tweet deleted.');
+            return redirect('/tweets')->with('success', 'Tweet deleted.');
+        }
+        // Redirect by default.
+        return redirect('/tweets');
     }
 }
